requirements.txt
groq
python-dotenv
pydantic
fastapi
uvicorn
